test(hibernate-reactive): use CompatibilityTestScenario enum to make interop test matrix explicit
Refactor the io.quarkus.hibernate.reactive.compatibility test suite to use a
CompatibilityTestScenario enum that defines all 10 ORM/Reactive interoperability
scenarios in one place. This makes the test coverage matrix explicit and eliminates
confusing test class names.

Each scenario is a named enum constant with Javadoc explaining which combination
of blocking/reactive/datasource/PU configuration it tests. Concrete test classes
are simplified: they reference the relevant scenario constant and focus only on
test methods and any scenario-specific CDI injection.

Old class names such as
ORMReactiveCompatibilityDifferentNamedDataSourceNamedPersistenceUnitBothUnitTest
are replaced with shorter, descriptive names like
DifferentNamedDatasourcesNamedPuBothCompatibilityTest.

Closes #51207
