--- extensions/hibernate-reactive/deployment/src/test/java/io/quarkus/hibernate/reactive/compatibility/ORMReactiveCompatibilityDefaultOnlyReactiveUnitTest.java
@@ -1,35 +1,0 @@
-package io.quarkus.hibernate.reactive.compatibility;
-
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.RegisterExtension;
-
-import io.quarkus.hibernate.reactive.entities.Hero;
-import io.quarkus.test.QuarkusExtensionTest;
-import io.quarkus.test.vertx.RunOnVertxContext;
-import io.quarkus.test.vertx.UniAsserter;
-
-public class ORMReactiveCompatibilityDefaultOnlyReactiveUnitTest extends CompatibilityUnitTestBase {
-
-    // To disable the blocking datasource, it's enough not to include the jdbc driver in the dependencies
-    @RegisterExtension
-    static final QuarkusExtensionTest config = new QuarkusExtensionTest()
-            .withApplicationRoot((jar) -> jar
-                    .addClasses(Hero.class)
-                    .addAsResource("complexMultilineImports.sql", "import.sql"))
-            .overrideConfigKey("quarkus.hibernate-orm.schema-management.strategy", SCHEMA_MANAGEMENT_STRATEGY)
-            .overrideConfigKey("quarkus.datasource.reactive", "true")
-            .overrideConfigKey("quarkus.datasource.db-kind", POSTGRES_KIND)
-            .overrideConfigKey("quarkus.datasource.username", USERNAME_PWD)
-            .overrideConfigKey("quarkus.datasource.password", USERNAME_PWD);
-
-    @Test
-    @RunOnVertxContext
-    public void testReactive(UniAsserter asserter) {
-        testReactiveWorks(asserter);
-    }
-
-    @Test
-    public void testBlocking() {
-        testBlockingDisabled();
-    }
-}
