--- extensions/hibernate-reactive/deployment/src/test/java/io/quarkus/hibernate/reactive/compatibility/ORMReactiveCompatibilityNamedDataSourceNamedPersistenceUnitBothUnitTest.java
@@ -1,56 +1,0 @@
-package io.quarkus.hibernate.reactive.compatibility;
-
-import java.util.List;
-
-import jakarta.inject.Inject;
-
-import org.hibernate.SessionFactory;
-import org.hibernate.reactive.mutiny.Mutiny;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.RegisterExtension;
-
-import io.quarkus.builder.Version;
-import io.quarkus.hibernate.orm.PersistenceUnit;
-import io.quarkus.hibernate.reactive.entities.Hero;
-import io.quarkus.maven.dependency.Dependency;
-import io.quarkus.test.QuarkusExtensionTest;
-import io.quarkus.test.vertx.RunOnVertxContext;
-import io.quarkus.test.vertx.UniAsserter;
-
-public class ORMReactiveCompatibilityNamedDataSourceNamedPersistenceUnitBothUnitTest extends CompatibilityUnitTestBase {
-
-    @RegisterExtension
-    static final QuarkusExtensionTest config = new QuarkusExtensionTest()
-            .withApplicationRoot((jar) -> jar
-                    .addClasses(Hero.class)
-                    .addAsResource("complexMultilineImports.sql", "import.sql"))
-            .setForcedDependencies(List.of(
-                    Dependency.of("io.quarkus", "quarkus-jdbc-postgresql-deployment", Version.getVersion()) // this triggers Agroal
-            ))
-            .overrideConfigKey("quarkus.hibernate-orm.\"named-pu\".schema-management.strategy", SCHEMA_MANAGEMENT_STRATEGY)
-            .overrideConfigKey("quarkus.hibernate-orm.\"named-pu\".datasource", "named-datasource")
-            .overrideConfigKey("quarkus.hibernate-orm.\"named-pu\".packages", "io.quarkus.hibernate.reactive.entities")
-            .overrideConfigKey("quarkus.datasource.\"named-datasource\".reactive", "true")
-            .overrideConfigKey("quarkus.datasource.\"named-datasource\".db-kind", POSTGRES_KIND)
-            .overrideConfigKey("quarkus.datasource.\"named-datasource\".username", USERNAME_PWD)
-            .overrideConfigKey("quarkus.datasource.\"named-datasource\".password", USERNAME_PWD)
-            .overrideConfigKey("quarkus.log.category.\"io.quarkus.hibernate\".level", "DEBUG");
-
-    @PersistenceUnit("named-pu")
-    Mutiny.SessionFactory namedMutinySessionFactory;
-
-    @Test
-    @RunOnVertxContext
-    public void test(UniAsserter uniAsserter) {
-        testReactiveWorks(namedMutinySessionFactory, uniAsserter);
-    }
-
-    @Inject
-    @PersistenceUnit("named-pu")
-    SessionFactory namedPersistenceUnitSessionFactory;
-
-    @Test
-    public void testBlocking() {
-        testBlockingWorks(namedPersistenceUnitSessionFactory);
-    }
-}
