--- extensions/hibernate-reactive/deployment/src/test/java/io/quarkus/hibernate/reactive/compatibility/ORMReactiveCompatibilityOnlyReactiveJDBCDisabledUnitTest.java
@@ -1,43 +1,0 @@
-package io.quarkus.hibernate.reactive.compatibility;
-
-import java.io.IOException;
-import java.util.List;
-
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.RegisterExtension;
-
-import io.quarkus.builder.Version;
-import io.quarkus.hibernate.reactive.entities.Hero;
-import io.quarkus.maven.dependency.Dependency;
-import io.quarkus.test.QuarkusExtensionTest;
-import io.quarkus.test.vertx.RunOnVertxContext;
-import io.quarkus.test.vertx.UniAsserter;
-
-public class ORMReactiveCompatibilityOnlyReactiveJDBCDisabledUnitTest extends CompatibilityUnitTestBase {
-
-    // We disable the JDBC data source witht the quarkus.datasource.jdbc=false we keep the driver
-    @RegisterExtension
-    static final QuarkusExtensionTest config = new QuarkusExtensionTest()
-            .withApplicationRoot((jar) -> jar
-                    .addClasses(Hero.class)
-                    .addAsResource("complexMultilineImports.sql", "import.sql"))
-            .setForcedDependencies(List.of(
-                    Dependency.of("io.quarkus", "quarkus-jdbc-postgresql-deployment", Version.getVersion())))
-            .overrideConfigKey("quarkus.hibernate-orm.schema-management.strategy", SCHEMA_MANAGEMENT_STRATEGY)
-            .overrideConfigKey("quarkus.datasource.jdbc", "false")
-            .overrideConfigKey("quarkus.datasource.reactive", "true")
-            .overrideConfigKey("quarkus.datasource.db-kind", POSTGRES_KIND)
-            .overrideConfigKey("quarkus.datasource.username", USERNAME_PWD)
-            .overrideConfigKey("quarkus.datasource.password", USERNAME_PWD);
-
-    @Test
-    @RunOnVertxContext
-    public void testReactive(UniAsserter asserter) {
-        testReactiveWorks(asserter);
-    }
-
-    @Test
-    public void testBlocking() throws IOException {
-        testBlockingDisabled();
-    }
-}
